add helptext, make JSON more complex to handle it
COMMAND() grows a fourth string, helptext; json now has a dict
for each opcode, with items keyed 'sig' with the descriptor array
and 'helptext' with the help text.  Perhaps the latter could grow
to a 'help' object with short and long help if needed.

Change all inclusions of MonCommands.h to reflect the new
four-param instances.